android release configuration - default icon and possible icons

+++ src/main/groovy/com/apphance/ameba/plugins/android/AndroidBuildXmlHelper.groovy
@@ -11,15 +11,20 @@ class AndroidBuildXmlHelper {
 
     String projectName(File projectDir) {
         def buildXml = new File(projectDir, BUILD_XML)
-        def xml = new XmlSlurper().parse(buildXml)
-        xml.@'name'.text()
+        if (buildXml?.exists()) {
+            def xml = new XmlSlurper().parse(buildXml)
+            return xml.@'name'.text()
+        }
+        ''
     }
 
     void replaceProjectName(File projectDir, String newName) {
         def buildXml = new File(projectDir, BUILD_XML)
-        def xml = new XmlSlurper().parse(buildXml)
-        xml.@'name' = newName
-        buildXml.delete()
-        buildXml.write(XmlUtil.serialize(xml))
+        if (buildXml?.exists()) {
+            def xml = new XmlSlurper().parse(buildXml)
+            xml.@'name' = newName
+            buildXml.delete()
+            buildXml.write(XmlUtil.serialize(xml))
+        }
     }
 }
